Fix 5 critical, 12 high, 6 medium audit findings across 39 files
Critical: remove hardcoded IBM Cloud CRN from runner.py, remove
fabricated CVE-2026-4539 from pip-audit, remove nonexistent tools/
COPY from Dockerfile, add SPDX headers to both pyproject.toml files.

High: add SPDX headers to 8 files (3 tests, 2 figures, scripts,
run_hardware, Cargo.toml), add line-6 descriptors to all 18
__init__.py files, replace noqa:F401 with __all__ re-export, fix
unnecessary type:ignore with np.float64, pin CI tools to ecosystem
versions (ruff==0.15.6, mypy==1.19.1, bandit==1.9.4).

Medium: refactor knm_to_hamiltonian as delegate to knm_to_xxz_hamiltonian
eliminating 30-line duplication, fix Makefile/pre-commit preflight
targets (tools/preflight.py never existed), update stale dates in
CITATION.cff and .zenodo.json, add lib.rs line-6 descriptor.

Also fixes 2 pre-existing test bugs: _R_from_xyz returns tuple not
scalar, _run_vqe dict keys are vqe_energy/exact_ground_energy not
ground_energy/exact_energy.

Verified: ruff clean, mypy 0 errors, 1755 tests pass (2 pre-existing
failures in test_rust_parity and test_ssgf_spectral unrelated).

+++ src/scpn_quantum_control/hardware/experiments.py
@@ -1130,8 +1130,8 @@ def decoherence_scaling_experiment(
             float(np.var(log_ratios)), 1e-10
         )
     else:
-        gamma = float("nan")  # type: ignore[assignment]
-        r_squared = float("nan")  # type: ignore[assignment]
+        gamma = np.float64("nan")
+        r_squared = np.float64("nan")
 
     print(f"  Fit: gamma={gamma:.6f} per gate, R²={r_squared:.4f}")
